More Refactoring
- Removed modifiers from the variables declared in ElkNodePainter constructor (makes everything immutable by default)

- Moved getElementForLocation from ElkNodePainter to JBlockDiagramVisualizer since it was not used for any painting functionality

- Moved edgeSectionPairs function to a new ElkNodeUtil object class
