+++ src/main/scala/edg_ide/swing/ElkNodePainter.scala
@@ -2,14 +2,16 @@ package edg_ide.swing
 
 import org.eclipse.elk.core.options._
 import org.eclipse.elk.graph._
-
+import edg_ide.swing.ElkNodeUtil.edgeSectionPairs
 import java.awt._
 import java.awt.geom.AffineTransform
 import scala.jdk.CollectionConverters.{ListHasAsScala, SetHasAsScala}
 
 
-class ElkNodePainter(var rootNode: ElkNode, var showTop: Boolean = false,
-                     var zoomLevel: Float = 1.0f, val margin: Int = 32) {
+class ElkNodePainter(rootNode: ElkNode, showTop: Boolean = false){
+  private val zoomLevel: Float = 1.0f
+  protected val margin: Int = 32
+
 
   def blendColor(baseColor: Color, topColor: Color, factor: Double): Color = {
     new Color(
@@ -54,84 +56,6 @@ class ElkNodePainter(var rootNode: ElkNode, var showTop: Boolean = false,
       base
     }
   }
-
-  // Utility functions
-  def edgeSectionPairs[T](section: ElkEdgeSection): Seq[((Double, Double), (Double, Double))] = {
-    val start = (section.getStartX, section.getStartY)
-    val end = (section.getEndX, section.getEndY)
-    val bends = section.getBendPoints.asScala.map { elkBendPoint =>
-      (elkBendPoint.getX, elkBendPoint.getY)
-    }
-    val allPoints = Seq(start) ++ bends ++ Seq(end)
-
-    allPoints.sliding(2).map { case point1 :: point2 :: Nil =>
-      (point1, point2)
-    }.toSeq
-  }
-
-  val EDGE_CLICK_WIDTH = 5.0f // how thick edges are for click detection purposes
-
-  def getElementForLocation(x: Int, y: Int): Option[ElkGraphElement] = {
-    def shapeContainsPoint(shape: ElkShape, point: (Double, Double)): Boolean = {
-      (shape.getX <= point._1 && point._1 <= shape.getX + shape.getWidth) &&
-        (shape.getY <= point._2 && point._2 <= shape.getY + shape.getHeight)
-    }
-
-    def lineClosestDist(line1: (Double, Double), line2: (Double, Double), point: (Double, Double)): Double = {
-      // Adapted from https://stackoverflow.com/questions/849211/shortest-distance-between-a-point-and-a-line-segment
-      val lengthSq = Math.pow(line2._1 - line1._1, 2) +
-        Math.pow(line2._2 - line1._2, 2)
-      if (lengthSq == 0) { // "line" is a point, return distance to the point
-        Math.sqrt(Math.pow(point._1 - line1._1, 2) + Math.pow(point._2 - line1._2, 2)).toFloat
-      } else {
-        val dot = (point._1 - line1._1) * (line2._1 - line1._1) +
-          (point._2 - line1._2) * (line2._2 - line1._2)
-        val t = Math.max(0, Math.min(1, dot / lengthSq))
-        val proj = (line1._1 + t * (line2._1 - line1._1),
-          line1._2 + t * (line2._2 - line1._2))
-        val dist = Math.sqrt(Math.pow(point._1 - proj._1, 2) + Math.pow(point._2 - proj._2, 2))
-        dist.toFloat
-      }
-    }
-
-    def edgeClosestDistance(edge: ElkEdge, point: (Double, Double)): Double = {
-      edge.getSections.asScala.map { section =>
-        edgeSectionPairs(section).map { case (line1, line2) =>
-          lineClosestDist(line1, line2, point)
-        }.min
-      }.min
-    }
-
-    // Tests the clicked point against a node, returning either a sub-node, port, or edge
-    def intersectNode(node: ElkNode, point: (Double, Double)): Option[ElkGraphElement] = {
-      // Ports can be outside the main shape and can't be gated by the node shape test
-      val nodePoint = (point._1 - node.getX, point._2 - node.getY) // transform to node space
-      val containedPorts = node.getPorts.asScala.collect {
-        case port if shapeContainsPoint(port, nodePoint) => port
-      }
-
-      // Test node, and if within node, recurse into children
-      val containedNodes = if (shapeContainsPoint(node, point)) {
-        val containedChildren = node.getChildren.asScala.flatMap(intersectNode(_, nodePoint))
-        val edgeDistances = node.getContainedEdges.asScala.map { edge =>
-          (edge, edgeClosestDistance(edge, nodePoint) * zoomLevel) // attach distance calculation
-        }.sortBy(_._2) // sort to get closest to cursor
-        val containedEdges = edgeDistances.collect { case (edge, dist)
-          if dist <= EDGE_CLICK_WIDTH => edge // filter by maximum click distance
-        }
-
-        containedChildren ++ containedEdges ++ Seq(node)
-      } else {
-        Seq()
-      }
-
-      (containedPorts ++ containedNodes).headOption
-    }
-
-    val elkPoint = ((x - margin) / zoomLevel.toDouble, (y - margin) / zoomLevel.toDouble) // transform points to elk-space
-    intersectNode(rootNode, elkPoint)
-  }
-
 
   // Given a ElkLabel and placement (anchoring) constraints, return the x and y coordinates for where the
   // text should be drawn.
